// database/supabase/types.ts
export type Json =
    | string
    | number
    | boolean
    | null
    | { [key: string]: Json | undefined }
    | Json[];

export type Database = {
    // Allows to automatically instanciate createClient with right options
    // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
    __InternalSupabase: {
        PostgrestVersion: "12.2.3 (519615d)";
    };
    public: {
        Tables: {
            backgrounds: {
                Row: {
                    background_config: Json;
                    created_at: string;
                    description: string | null;
                    id: number;
                    is_featured: boolean | null;
                    name: string;
                    sort_order: number | null;
                };
                Insert: {
                    background_config: Json;
                    created_at?: string;
                    description?: string | null;
                    id?: number;
                    is_featured?: boolean | null;
                    name: string;
                    sort_order?: number | null;
                };
                Update: {
                    background_config?: Json;
                    created_at?: string;
                    description?: string | null;
                    id?: number;
                    is_featured?: boolean | null;
                    name?: string;
                    sort_order?: number | null;
                };
                Relationships: [];
            };
            default_geometries: {
                Row: {
                    color_day: string | null;
                    color_night: string | null;
                    created_at: string;
                    description: string | null;
                    geometry_type: string;
                    id: number;
                    is_featured: boolean | null;
                    material_config: Json | null;
                    name: string;
                    sort_order: number | null;
                };
                Insert: {
                    color_day?: string | null;
                    color_night?: string | null;
                    created_at?: string;
                    description?: string | null;
                    geometry_type: string;
                    id?: number;
                    is_featured?: boolean | null;
                    material_config?: Json | null;
                    name: string;
                    sort_order?: number | null;
                };
                Update: {
                    color_day?: string | null;
                    color_night?: string | null;
                    created_at?: string;
                    description?: string | null;
                    geometry_type?: string;
                    id?: number;
                    is_featured?: boolean | null;
                    material_config?: Json | null;
                    name?: string;
                    sort_order?: number | null;
                };
                Relationships: [];
            };
// ...
        };
        Views: {
            [_ in never]: never;
        };
        Functions: {
            [_ in never]: never;
        };
        Enums: {
            [_ in never]: never;
        };
        CompositeTypes: {
            [_ in never]: never;
        };
    };
};